MAINT: generalised spectral models and moved utils from plotting
