+++ matador/orm/spectral/__init__.py
@@ -3,12 +3,15 @@ DOS and dispersion data.
 
 """
 
-from .dos import VibrationalDOS, ElectronicDOS
-from .dispersion import VibrationalDispersion, ElectronicDispersion
+from .dos import VibrationalDOS, ElectronicDOS, DensityOfStates
+from .dispersion import VibrationalDispersion, ElectronicDispersion, Dispersion
+
 
 __all__ = [
     'VibrationalDOS',
     'ElectronicDOS',
+    'DensityOfStates',
     'VibrationalDispersion',
-    'ElectronicDispersion'
+    'ElectronicDispersion',
+    'Dispersion'
 ]
